+++ entrypoint.sh
@@ -3,7 +3,6 @@
 RUNTIME_ENV_CONFIG="
     window.srConfig = {};
     window.srConfig.API_URL = \"$API_URL\";
-    window.srConfig.MEILISEARCH_HOST = \"$MEILISEARCH_HOST\";
 "
 echo $RUNTIME_ENV_CONFIG > /app/dist/config.js