+++ src/components/Cocktail/IngredientModal.vue
@@ -1,13 +1,18 @@
 <script setup>
 import Modal from './../Modal.vue'
+import OverlayLoader from './../OverlayLoader.vue'
 </script>
 
 <template>
-    <Modal>
+    <Modal class="ingredient-modal">
         <template #body>
+            <OverlayLoader v-if="isLoading" />
             <ais-instant-search :search-client="searchClient" :index-name="index" :on-state-change="onStateChange">
                 <ais-configure :hitsPerPage="30" />
                 <ais-search-box placeholder="Search for ingredient..." :class-names="{'ais-SearchBox-input': 'form-input'}" />
+                <label for="substitute-adding">
+                    <input id="substitute-adding" type="checkbox" v-model="isAddingSubstitute"> Select substitute ingredients
+                </label>
                 <ais-hits>
                     <template v-slot="{ items }">
                         <div class="ingredients-options">
@@ -23,6 +28,10 @@ import Modal from './../Modal.vue'
                 <small>Current ingredient:</small>
                 {{ cocktailIngredient.name }}
             </h3>
+            <div class="substitutes">
+                <small>Substitutes:</small>
+                <span v-for="substitute in cocktailIngredient.substitutes">{{ substitute.name }} &middot; <a href="#" @click.prevent="removeSubstitute(substitute)">Remove</a></span>
+            </div>
             <div style="display: grid; grid-template-columns: 1fr 2fr; column-gap: 10px;">
                 <div class="form-group">
                     <label class="form-label" for="ingredient-amount">Amount:</label>
@@ -42,7 +51,7 @@ import Modal from './../Modal.vue'
         </template>
         <template #footer>
             <button type="button" class="button button--outline" @click="cancel" style="margin-right: 10px;">Cancel</button>
-            <button type="button" class="button button--dark" @click="save">Done</button>
+            <button type="button" class="button button--dark" @click="save" :disabled="isLoading">Done</button>
         </template>
     </Modal>
 </template>
@@ -67,6 +76,11 @@ export default {
             // Reset search on modal open
             document.querySelector('.modal-body form').reset()
 
+            // Add substitutes array
+            if (!this.cocktailIngredient.substitutes) {
+                this.cocktailIngredient.substitutes = []
+            }
+
             // Focus seach input on modal open
             setTimeout(() => {
                 document.querySelector('.ais-SearchBox input').focus()
@@ -75,6 +89,7 @@ export default {
     },
     data() {
         return {
+            isLoading: false,
             index: 'ingredients:name:asc',
             searchClient: instantMeiliSearch(
                 Auth.getUserSearchSettings().host,
@@ -83,23 +98,39 @@ export default {
             cocktailIngredient: {},
             orgCocktailIngredient: {},
             currentQuery: null,
+            isAddingSubstitute: false
         }
     },
     methods: {
         selectIngredient(item) {
-            this.cocktailIngredient.ingredient_id = item.id;
-            this.cocktailIngredient.name = item.name;
-            this.cocktailIngredient.ingredient_slug = item.slug;
+            if (this.isAddingSubstitute) {
+                if (this.cocktailIngredient.substitutes && this.cocktailIngredient.substitutes.some(sub => sub.id == item.id)) {
+                    return;
+                }
+
+                this.cocktailIngredient.substitutes.push({
+                    id: item.id,
+                    name: item.name,
+                    slug: item.slug,
+                })
+            } else {
+                this.cocktailIngredient.ingredient_id = item.id;
+                this.cocktailIngredient.name = item.name;
+                this.cocktailIngredient.ingredient_slug = item.slug;
+            }
         },
         save() {
             window.document.body.style.overflow = 'auto';
+            this.isAddingSubstitute = false;
             this.$emit('close');
         },
         cancel() {
             window.document.body.style.overflow = 'auto';
+            this.isAddingSubstitute = false;
             this.$emit('close');
         },
         newIngredient() {
+            this.isLoading = true;
             ApiRequests.saveIngredient({
                 name: this.currentQuery,
                 description: null,
@@ -115,9 +146,17 @@ export default {
                     slug: data.slug,
                     id: data.id
                 });
+                this.isLoading = false;
             }).catch(() => {
                 this.$toast.error('Unable to add ingredient.');
+                this.isLoading = false;
             })
+        },
+        removeSubstitute(ing) {
+            this.cocktailIngredient.substitutes.splice(
+                this.cocktailIngredient.substitutes.findIndex(i => i == ing),
+                1
+            );
         },
         onStateChange({ uiState, setUiState }) {
             const indexState = uiState[this.index] || {}
@@ -170,18 +209,32 @@ export default {
     line-height: 1.3;
 }
 
-.selected-ingredient small {
+.selected-ingredient small,
+.ingredient-modal .substitutes small {
     font-weight: normal;
     font-size: 0.8rem;
     display: block;
     color: #866269;
 }
 
+.ingredient-modal .substitutes {
+    margin-bottom: 20px;
+}
+
+.ingredient-modal .substitutes span {
+    font-weight: normal;
+    font-size: 1rem;
+    display: block;
+    font-family: var(--font-accent);
+    font-weight: 700;
+}
+
 .ingredient-modal__info {
-    background-color: rgb(228, 241, 252);
+    background: rgba(255, 255, 255, .5);
+    border-bottom: 2px solid var(--color-bg-dark);
+    border-radius: 5px;
     color: rgb(28, 48, 65);
     padding: 10px;
-    border-radius: 5px;
     margin-top: 10px;
 }
 </style>