+++ src/components/Ingredient/Form.vue
@@ -1,3 +1,8 @@
+<script setup>
+import ImageUpload from './../ImageUpload.vue'
+import { ColorPicker } from 'vue-accessible-color-picker'
+</script>
+
 <template>
     <form @submit.prevent="submit">
         <h2 class="page-subtitle">Ingredient information</h2>
@@ -26,16 +31,12 @@
         </div>
         <div class="form-group">
             <label class="form-label" for="color">Color:</label>
-            <input class="form-input" type="text" id="color" v-model="ingredient.color">
+            <button type="button" class="colorpicker-button" @click="showColorPicker = !showColorPicker">
+                <span :style="{'background-color': ingredient.color}"></span>
+            </button>
+            <ColorPicker v-if="showColorPicker" alpha-channel="hide" :visible-formats="['hex']" :color="ingredient.color" @color-change="updateColor" />
         </div>
-        <div class="form-group">
-            <label class="form-label" for="images">Images:</label>
-            <input class="form-input" type="file" id="images" ref="image">
-        </div>
-        <div class="form-group">
-            <label class="form-label" for="copyright">Image copyright:</label>
-            <input class="form-input" type="text" id="copyright" v-model="images[0].copyright">
-        </div>
+        <ImageUpload ref="imagesUpload" :value="ingredient.images" />
         <div class="form-actions">
             <RouterLink class="button button--outline" :to="{name: 'ingredients'}">Cancel</RouterLink>
             <button class="button button--dark" type="submit">Save</button>
@@ -50,6 +51,7 @@ export default {
     data() {
         return {
             ingredientId: null,
+            showColorPicker: false,
             ingredient: {},
             images: [
                 {image: null, copyright: null}
@@ -87,21 +89,11 @@ export default {
                 ingredient_category_id: this.ingredient.ingredient_category_id,
             };
 
-            const image = this.$refs.image.files[0] || null;
-
-            if (image) {
-                const formData = new FormData();
-                formData.append('images[0][image]', image)
-                formData.append('images[0][copyright]', this.images[0].copyright)
-
-                const resp = await ApiRequests.uploadImages(formData).catch(e => {
-                    this.$toast.error('An error occured while uploading images. Your ingredient is still saved.');
-                });
-
-                if (resp) {
-                    postData.images.push(resp[0].id);
-                }
-            }
+            const imageResources = await this.$refs.imagesUpload.uploadPictures().catch(() => {
+                this.$toast.error('An error occured while uploading images. Your ingredient is still saved.');
+            }) || [];
+            
+            postData.images = imageResources.map(img => img.id);
 
             if (this.ingredientId) {
                 ApiRequests.updateIngredient(this.ingredientId, postData).then(data => {
@@ -132,7 +124,37 @@ export default {
                     this.isLoading = false;
                 })
             }
+        },
+        updateColor(eventData) {
+            this.ingredient.color = eventData.cssColor
         }
     }
 }
 </script>
+<style scoped>
+.colorpicker-button {
+    cursor: pointer;
+    padding: 10px;
+    width: 100%;
+    display: flex;
+    background: rgba(255, 255, 255, .5);
+    border-radius: 5px;
+    height: 3rem;
+    border: 2px solid var(--color-bg-dark);
+    border-top-color: transparent;
+    border-left-color: transparent;
+    border-right-color: transparent;
+}
+
+.colorpicker-button span {
+    display: flex;
+    width: 100%;
+}
+
+.colorpicker-button:hover,
+.colorpicker-button:active,
+.colorpicker-button:focus {
+    background: #fff;
+    border-color: var(--color-text);
+}
+</style>