+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# v0.3.0
+## New
+- Support adding cocktail ingredient substitutes
+- Show ingredient varieties
+- Add "cl" as unit of measurement
+- New image uploading component
+- Add glass types updating
+
+## Misc
+- Meilisearch host server is now fetched from user endpoint
+- Update cocktail card styling
+- Truncate a lot of tags on cocktail card
+- Missing cocktail and ingredient images are now local to client
+- Add colorpicker to ingredient form
+
 # v0.2.1
 - Some fixes for ingredient modal
 - Fix image upload container on smaller devices